Everything except DEL works properly

+++ Makefile
@@ -12,7 +12,6 @@ all: forests
 
 $(TEST_EXECUTABLE): CFLAGS += -pedantic -Werror -fstack-protector-strong -g
 $(TEST_EXECUTABLE): $(SRCDIR)/forests.c $(BUILDDIR)/bst.o
-	mkdir -p $(BUILDDIR)
 	$(CC) $(CFLAGS) -c $< -o $(BUILDDIR)/forest.o
 	$(CC) -o $@ $(BUILDDIR)/forest.o $(BUILDDIR)/bst.o
 
@@ -21,11 +20,12 @@ forests: $(SRCDIR)/forests.c $(BUILDDIR)/bst.o
 	$(CC) -o $@ $(BUILDDIR)/forest.o $(BUILDDIR)/bst.o
 
 $(BUILDDIR)/bst.o: $(SRCDIR)/bst.c $(SRCDIR)/bst.h
+	mkdir -p $(BUILDDIR)
 	$(CC) $(CFLAGS) -c $< -o $@
 
 clean:
-	rm -rf build
-	rm forests
+	if [ -d "$(BUILDDIR)" ]; then rm -rf "$(BUILDDIR)"; fi
+	if [ -f forests ]; then rm forests; fi
 
 test: test.sh $(TEST_EXECUTABLE) $(TESTSDIR)
 	./test.sh $(TEST_EXECUTABLE) $(TESTSDIR)
